requirements.txt
fastapi
uvicorn
pydantic
python-multipart
sqlalchemy
faiss-cpu
sentence-transformers
ultralytics
openai
